image null 허용, ModelAndView를 내려주는 MainController 생성, 모든 요청에 user 정보 내려주기

+++ domain/src/main/java/sion/bookstore/domain/auth/User.java
@@ -2,7 +2,7 @@ package sion.bookstore.domain.auth;
 
 public interface User {
     boolean authenticated();
-    Integer getMemberId();
+    Long getMemberId();
     String getUserEmail();
     String getUserName();
     String getAccessIp();
